Fix extract-data logic - return when TOML decoding fails
Currently in `extract_data` if `toml.load` fails data is unset. Hence
subsequent calls to data['name'] will fail. The correct approach is to
return immediately in such cases.

Also TomlDecodeError may be triggered due to `toml.load`, so that's the
only directive that must be wrapped in a `try-except` statement.