Update linter tests based on established colcon templates (#34)
In addition to aligning this package's linters with the other supported
colcon extensions, there was a bug in the spelling linter and it wasn't
actually working. This change also fixes that, and the handful of
misspellings that had slipped through.

I pretty much overwrote the linter test files with ones copied from
colcon/template-package and then replaced the directory names to match
colcon-cargo.