Add '+upstream' suffix to published deb version
We continue to see interference between the deb packages we publish from
the colcon project and efforts to package colcon as part of mainline
Debian and Ubuntu. Using a high Debian-Version value mitigated the
problems in most cases, but was not sufficient to eliminate all of the
conflicts we're currently experiencing.

Using a debian version suffix which falls late alphabetically appears to
give our packages preference by apt. If a user enables a repository
which distributes packages created by the colcon project, it is likely
that they wish to use these packages instead of the ones packaged by
their platform.