Add support for cargo dependency renaming (#55)
It is possible to reference a crate using an import name other than its
package name. The package name and import name are typically the same,
but when they differ, we should use the package name to construct the
dependency graph. The import name is of no consequence outside of the
package being built.