Extract {build,dev}-dependencies
The build-dependencies are needed by build scripts, which makes it a
sort of "export" or "devel" dependency of the package. Since colcon
doesn't currently disambiguate between this class and "run"
dependencies, we will need to include the dependencies under "run".

The "dev" dependencies however are not needed for dependees of a
package, and only the package itself.