Specify package name in cargo operations
When inspecting the output from `cargo metadata` and when invoking other
cargo commands like `build`, `test`, and `fmt`, we should specify the
name of the package we're operating on.

When a package is part of a cargo workspace, these operations will
normally encompass the whole workspace but colcon operates on a
per-package basis, so we need to limit the scope to the specific package
we're working with.