Add support for cargo workspaces
This change adds support for cargo workspaces, both virtual workspaces
and ones with top-level packages. It does so by caching the paths to the
workspace members as the workspaces are discovered by "primary"
discovery mechanisms (like a recursive crawl) and then specifically
loads packages from those directories after the primary discovery
extensions have completed.

Note that all of the packages in a workspace are discovered regardless
of the "default-members" value in the workspace configuration.

A subtle change in behavior introduced by this change is that any
packages in subdirectories of a workspace which are not listed as a
workspace member will not be discovered. The presence of a valid
workspace configuration, while not strictly a package itself, should
serve as a declaration of what packages live in subdirectories of said
workspace and colcon should respect that.