Add support for cargo workspaces (#59)
* Specify package name in cargo operations

When inspecting the output from `cargo metadata` and when invoking other
cargo commands like `build`, `test`, and `fmt`, we should specify the
name of the package we're operating on.

When a package is part of a cargo workspace, these operations will
normally encompass the whole workspace but colcon operates on a
per-package basis, so we need to limit the scope to the specific package
we're working with.

* Add support for cargo workspaces

This change adds support for cargo workspaces, both virtual workspaces
and ones with top-level packages. It does so by caching the paths to the
workspace members as the workspaces are discovered by "primary"
discovery mechanisms (like a recursive crawl) and then specifically
loads packages from those directories after the primary discovery
extensions have completed.

Note that all of the packages in a workspace are discovered regardless
of the "default-members" value in the workspace configuration.

A subtle change in behavior introduced by this change is that any
packages in subdirectories of a workspace which are not listed as a
workspace member will not be discovered. The presence of a valid
workspace configuration, while not strictly a package itself, should
serve as a declaration of what packages live in subdirectories of said
workspace and colcon should respect that.

* Use a separate identification extension for workspaces

* Filter out false detection of circular dependencies (#62)

* Add a comment about ignoring other workspace members

Co-authored-by: Michael X. Grey <mxgrey@intrinsic.ai>