Allow specification of additional packages in workspaces (#66)
When cargo itself encounters a workspace, it inspects member directories
for crates and does not search any other subdirectories of the workspace
for packages. The workspace support in colcon behaves the same.

It may be desirable to include additional packages in colcon's discovery
which are not members of the workspace. This change adds a metadata
value at `workspace.metadata.colcon.additional-packages` which behaves
similarly to `workspace.members` so that packages under those paths are
discovered by colcon. The packages need not be cargo packages themselves
but if they are, they will not be treated as cargo workspace members but
rather as independent packages.